refactor: extract map pool into src/maps/ module
Move src/queue/map-pool/, apply-map-cooldown.ts, and reset-map-options.ts
to a new src/maps/ module shared between auto and captain queue modes.
No logic changes.
