+++ src/maps/apply-cooldown.ts
